# zstd_stream.go: remove intermediate source buffer for Writer.