The Writer was using an intermediate source buffer, which could sometimes buffer input bytes until the next Write, Flush, or Close call. This is not necessary, complicates the code, and can cause a performance hit.

This commit eliminates the source buffer and its overhead, instead ensuring that all input bytes are compressed when Write returns.

In practice the intermediate source buffer should never or almost never have been used, so this is mostly beneficial for simplifying the code, although we do see a memory usage improvement in the benchmarks

```
name                    old time/op    new time/op    delta
StreamCompression-16      57.9ms ± 5%    56.7ms ± 5%    ~     (p=0.165 n=10+10)

name                    old speed      new speed      delta
StreamCompression-16    1.81GB/s ± 5%  1.85GB/s ± 5%    ~     (p=0.165 n=10+10)

name                    old alloc/op   new alloc/op   delta
StreamCompression-16      2.19MB ± 7%    2.07MB ± 6%  -5.45%  (p=0.001 n=10+10)

name                    old allocs/op  new allocs/op  delta
StreamCompression-16        3.00 ± 0%      3.00 ± 0%    ~     (all equal)
```