@knusbaum 2024-10-02
Fixed. I'll merge this once CI is done.